+++ package.json
@@ -12,6 +12,7 @@
     "body-parser": "^1.17.1",
     "express": "^4.15.2",
     "express-rate-limit": "^2.6.0",
+    "express-routes-versioning": "^1.0.0",
     "request": "^2.81.0"
   },
   "devDependencies": {
