+++ server.js
@@ -12,6 +12,7 @@ var limiter = new RateLimit({
 });
 
 //  apply to all requests
+app.set('json spaces', 2);
 app.use(limiter);
 app.use(bodyParser.urlencoded({ extended: true }));
